testing

+++ src/NorthWestCornerMethod.java
@@ -46,7 +46,7 @@ public class NorthWestCornerMethod {
         int i = 0;
         int j = 0;
         while (i < solution.getRows() && j < solution.getRows()) {
-            int allocation = min(demandCoefficients.getElement(j, 0), supplyCoefficients.getElement(i, 0));
+            int allocation = (int) min(demandCoefficients.getElement(j, 0), supplyCoefficients.getElement(i, 0));
             demand.setElement(j, 0, demand.getElement(j, 0) - allocation);
             supply.setElement(i, 0, supply.getElement(i, 0) - allocation);
